Rename Instance class and update namespace to Di\Definition; modify exception message for empty component ID.
