+++ src/Factory/Exception/Message.php
@@ -12,7 +12,6 @@ namespace PHPPress\Factory\Exception;
  */
 enum Message: string
 {
-    case COMPONENT_ID_EMPTY = 'The required component "id" is empty.';
     case DEFINITION_EMPTY = 'The definition is empty.';
     case DEFINITION_INVALID = 'Invalid definition for "%s": %s';
     case DEFINITION_REQUIRES_CLASS_OPTION = 'A class definition requires a "__class" or "class" member.';
@@ -26,7 +25,6 @@ enum Message: string
     public function getMessage(string ...$argument): string
     {
         return match ($this) {
-            self::COMPONENT_ID_EMPTY,
             self::DEFINITION_EMPTY,
             self::DEFINITION_REQUIRES_CLASS_OPTION,
             self::DEPENDENCIES_IDX_NAME_POSITION => $this->value,
