+++ tests/Di/ContainerTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PHPPress\Tests\Di;
 
 use PHPPress\Di\Container;
-use PHPPress\Factory\Instance;
+use PHPPress\Di\Definition\Instance;
 use PHPPress\Tests\Provider\ContainerProvider;
 use PHPUnit\Framework\Attributes\{DataProviderExternal, Group};
