+++ tests/Di/InvokeableTest.php
@@ -6,11 +6,11 @@ namespace PHPPress\Tests\Di;
 
 use ArrayIterator;
 use DateTime;
-use PHPPress\Exception\{InvalidArgument, InvalidDefinition};
-use Psr\Container\ContainerInterface;
 use PHPPress\Di\Container;
-use PHPPress\Factory\Instance;
+use PHPPress\Di\Definition\Instance;
+use PHPPress\Exception\{InvalidArgument, InvalidDefinition};
 use PHPUnit\Framework\Attributes\Group;
+use Psr\Container\ContainerInterface;
 
 /**
  * Test cases for invokeable class handling in the dependency injection container.
